Refactor classes and packages, update style in edit text
Refactor classes and packages to corresponding locations, update style in edit text to be inline with the rest of the GUI

Issues:
#79

Contributors:
@krasen86
@chrytsa
@samgun-6
@eemilj
