+++ app/src/main/res/layout/car_select.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui_activity_controller.CarSelect">
+    tools:context=".ui_activity_controller.car_selection.CarSelect">
 
     <TextView
         android:id="@+id/chooseCarText"
