+++ app/src/main/res/layout/edit_popup.xml
@@ -3,6 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="400dp"
     android:layout_height="200dp"
+    android:layout_gravity="center_horizontal"
     style="@style/standard_list_item"
     android:orientation="vertical"
     >
@@ -22,10 +23,12 @@
             android:layout_width ="match_parent"
             android:layout_height="wrap_content"
             android:layout_gravity="center_vertical"
-            android:paddingStart="4dp"
+            android:paddingStart="5dp"
+            android:layout_marginStart="10dp"
+            android:layout_marginEnd="10dp"
             android:layout_marginTop="50dp"
-            android:layout_marginRight="23dp"
             android:autofillHints=""
+            android:textColor="@color/colorDeepJungleGreen"
             android:inputType="text"
             tools:ignore="LabelFor,RtlSymmetry"
             />
